requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-auth-ldap==5.1.0
django-role-permissions==3.2.0
fonttools==4.55.0
gunicorn==23.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
mysqlclient==2.2.6
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip-review==1.3.0
plotly==5.24.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-ldap==3.4.4
pytz==2024.2
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.2
tenacity==9.0.0
tzdata==2024.2
